start of ex01 + ex00 need te be tested a little more

+++ ex00/inc/BitcoinExchange.hpp
@@ -42,7 +42,7 @@ class BitcoinExchange {
         bool        checkRequirements(long int year, long int month, long int day,
             std::string &error_message, int &err);
         long int    extractDateValue(const std::string &date, int &err, std::string &error_message, bool csv);
-        long int    returnDataValue(const std::string &input_date, int &err,
+        long double    returnDataValue(const std::string &input_date, int &err,
             std::string &error_message);
         long int calculTheExchange(const std::string &s1, const std::string &s2,
             std::string &error_message);
